+++ packages/client/hmi-client/src/components/workflow/ops/simulate-ensemble-ciemss/mod.ts
@@ -1,6 +1,6 @@
 import { SimulateEnsembleCiemssOperation as operation } from './simulate-ensemble-ciemss-operation';
 import node from './tera-simulate-ensemble-node-ciemss.vue';
-import drilldown from './tera-simulate-ensemble-ciemss.vue';
+import drilldown from './tera-simulate-ensemble-ciemss-drilldown.vue';
 
 const name = operation.name;
