+++ packages/client/hmi-client/src/page/WorkflowNode.vue
@@ -8,7 +8,10 @@
 		<tera-calibrate-ciemss v-else-if="isNodeType(OperationType.CALIBRATION_CIEMSS)" :node="node" />
 		<tera-calibrate-ensemble-ciemss v-else-if="isNodeType(OperationType.CALIBRATE_ENSEMBLE_CIEMSS)" :node="node" />
 		<tera-simulate-ciemss-drilldown v-else-if="isNodeType(OperationType.SIMULATE_CIEMSS)" :node="node" />
-		<tera-simulate-ensemble-ciemss v-else-if="isNodeType(OperationType.SIMULATE_ENSEMBLE_CIEMSS)" :node="node" />
+		<tera-simulate-ensemble-ciemss-drilldown
+			v-else-if="isNodeType(OperationType.SIMULATE_ENSEMBLE_CIEMSS)"
+			:node="node"
+		/>
 		<tera-funman v-else-if="isNodeType(OperationType.FUNMAN)" :node="node" />
 		<tera-code-asset-wrapper v-else-if="isNodeType(OperationType.CODE)" :node="node" />
 	</template>
@@ -27,7 +30,7 @@ import TeraDatasetTransformer from '@/components/workflow/ops/dataset-transforme
 import TeraCalibrateCiemss from '@/components/workflow/ops/calibrate-ciemss/tera-calibrate-ciemss-drilldown.vue';
 import TeraCalibrateEnsembleCiemss from '@/components/workflow/ops/calibrate-ensemble-ciemss/tera-calibrate-ensemble-ciemss.vue';
 import TeraSimulateCiemssDrilldown from '@/components/workflow/ops/simulate-ciemss/tera-simulate-ciemss-drilldown.vue';
-import TeraSimulateEnsembleCiemss from '@/components/workflow/ops/simulate-ensemble-ciemss/tera-simulate-ensemble-ciemss.vue';
+import TeraSimulateEnsembleCiemssDrilldown from '@/components/workflow/ops/simulate-ensemble-ciemss/tera-simulate-ensemble-ciemss-drilldown.vue';
 import TeraFunman from '@/components/workflow/ops/funman/tera-funman.vue';
 import TeraStratifyMira from '@/components/workflow/ops/stratify-mira/tera-stratify-mira.vue';
 import TeraCodeAssetWrapper from '@/components/workflow/ops/code-asset/tera-code-asset-wrapper.vue';
